Inittial Creation
Created a Python app that will read a list from a .txt file named "Dictionary.txt" that will export it into a file called NewDictionary.txt". It automatically formats the list so it is easier to put into an array in several languages by adding "" arround the word and a comma after each word.